At lease use SOME security
UUIDs are not secure enough to use as tokens. Having said that, I see no reason to make this issue worse with version 1 UUIDs.

If you are using this code as an example of how to impliment the API interface to my hijacking prevention system, DON'T.  This example is not secure, even with minor fixes like this.